move InputPropertiesError back to its own package
