+++ pkg/resource/deploy/source_eval_test.go
@@ -3379,7 +3379,7 @@ func TestValidationFailures(t *testing.T) {
 		{
 			name: "bad request",
 			err:  badRequestError,
-			expectedStderr: "error: pulumi:providers:some-type resource 'some-name' has a problem: bad request:\n" +
+			expectedStderr: "error: pulumi:providers:some-type resource 'some-name' has a problem: bad request\n" +
 				"\t\t- property testproperty with value '{testvalue}' has a problem: missing\n" +
 				"\t\t- property nested[0] with value '{nestedvalue}' has a problem: nested property error\n",
 		},
