+++ sdk/go/common/util/rpcutil/rpcerror/rpcerror.go
@@ -37,6 +37,7 @@ import (
 	"google.golang.org/protobuf/runtime/protoiface"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
+	perrors "github.com/pulumi/pulumi/sdk/v3/go/pulumi/errors"
 	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
 )
 
@@ -47,7 +48,7 @@ type Error struct {
 	code                  codes.Code
 	message               string
 	cause                 *ErrorCause
-	inputPropertiesErrors []InputPropertyErrorDetails
+	inputPropertiesErrors []perrors.InputPropertyErrorDetails
 	details               []interface{}
 }
 
@@ -86,7 +87,7 @@ func (r *Error) Details() []interface{} {
 
 // InputPropertiesErrors returns the list of input properties error that
 // were attached to this error.
-func (r *Error) InputPropertiesErrors() []InputPropertyErrorDetails {
+func (r *Error) InputPropertiesErrors() []perrors.InputPropertyErrorDetails {
 	return r.inputPropertiesErrors
 }
 
@@ -154,7 +155,7 @@ func Wrapf(code codes.Code, err error, messageFormat string, args ...interface{}
 }
 
 func WrapDetailedError(err error) error {
-	var iperr *InputPropertiesError
+	var iperr *perrors.InputPropertiesError
 	if errors.As(err, &iperr) {
 		status := status.New(codes.InvalidArgument, iperr.Message)
 		errorDetails := pulumirpc.InputPropertiesError{}
@@ -204,9 +205,9 @@ func FromError(err error) (*Error, bool) {
 	rpcError.details = status.Details()
 	for _, details := range status.Details() {
 		if d, ok := details.(*pulumirpc.InputPropertiesError); ok {
-			rpcError.inputPropertiesErrors = make([]InputPropertyErrorDetails, len(d.Errors))
+			rpcError.inputPropertiesErrors = make([]perrors.InputPropertyErrorDetails, len(d.Errors))
 			for i, e := range d.GetErrors() {
-				rpcError.inputPropertiesErrors[i] = InputPropertyErrorDetails{
+				rpcError.inputPropertiesErrors[i] = perrors.InputPropertyErrorDetails{
 					PropertyPath: e.GetPropertyPath(),
 					Reason:       e.GetReason(),
 				}
@@ -252,70 +253,4 @@ func serializeErrorCause(err error) *pulumirpc.ErrorCause {
 		Message:    message,
 		StackTrace: stackTrace,
 	}
-}
-
-// InputPropertyErrorDetails contains the error details for an input property error.
-type InputPropertyErrorDetails struct {
-	PropertyPath string
-	Reason       string
-}
-
-func (d InputPropertyErrorDetails) String() string {
-	return fmt.Sprintf("%s: %s", d.PropertyPath, d.Reason)
-}
-
-// InputPropertiesError can be used to indicate that the client has made a request with
-// bad input properties.
-type InputPropertiesError struct {
-	Message string
-	Errors  []InputPropertyErrorDetails
-}
-
-// Create a new InputPropertiesError with a single property error.
-func NewInputPropertyError(propertyPath string, reason string) *InputPropertiesError {
-	return NewInputPropertiesError("", InputPropertyErrorDetails{
-		PropertyPath: propertyPath,
-		Reason:       reason,
-	})
-}
-
-// Create a new InputPropertiesError with a single property error.
-func InputPropertyErrorf(propertyPath string, format string, args ...interface{}) *InputPropertiesError {
-	return NewInputPropertiesError("", InputPropertyErrorDetails{
-		PropertyPath: propertyPath,
-		Reason:       fmt.Sprintf(format, args...),
-	})
-}
-
-// Create a new InputPropertiesError with a message and a list of property errors.
-func NewInputPropertiesError(message string, details ...InputPropertyErrorDetails) *InputPropertiesError {
-	return &InputPropertiesError{
-		Message: message,
-		Errors:  details,
-	}
-}
-
-// Create a new InputPropertiesError with a message.
-func InputPropertiesErrorf(format string, args ...interface{}) *InputPropertiesError {
-	return NewInputPropertiesError(fmt.Sprintf(format, args...))
-}
-
-func (ipe *InputPropertiesError) Error() string {
-	message := ipe.Message
-	if message != "" && len(ipe.Errors) > 0 {
-		message += ": "
-	}
-	for i, err := range ipe.Errors {
-		if i == 0 {
-			message += "\n "
-		}
-		message += err.String()
-	}
-	return message
-}
-
-// WithDetails adds additional property errors to an existing InputPropertiesError.
-func (ipe *InputPropertiesError) WithDetails(details ...InputPropertyErrorDetails) *InputPropertiesError {
-	ipe.Errors = append(ipe.Errors, details...)
-	return ipe
 }
